refactor: refactor buffer_pool.rs into multiple files
