+++ crates/storage/src/buffer_pool/mod.rs
@@ -1,0 +1,9 @@
+pub mod manager;
+pub mod replacer;
+pub mod shard;
+
+#[cfg(test)]
+mod tests;
+
+pub use manager::{BufferPoolError, BufferPoolManager, Result};
+pub use shard::{BufferPoolShard, PageReadGuard, PageWriteGuard};
